+++ src/api/endpoints/__init__.py
@@ -1,2 +1,3 @@
 """Exposed endpoints classes and methods."""
 from api.endpoints.user_endpoint import user_router
+from api.endpoints.group_endpoint import group_router