+++ CHANGELOG.md
@@ -5,5 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ## Added
+- @javierfh03 - Develop user endpoint.
+- @javierfh03 - Develop group endpoint.
 
 [Unreleased]: https://github.com/Lagatrix/lagatrix_api